add resistance from dsx

+++ Utils/ByteUtil.cs
@@ -2,7 +2,7 @@
 
 internal class ByteUtil
 {
-    public static byte OffsetByte(int value, byte offset)
+    public static byte OffsetByte(ulong value, byte offset)
     {
         return (byte)((value >> offset) & 0xFF);
     }
